Hashing the passwords
Hashing the passwords with sha256 - build in algorithm with PHP. We create a hashed password for the user as they create an account, and then when they try to log in we rehash their inputted password. If the passwords are the same, then the hashes will come out the same, and users are able to have a simple password like '123' with the hash still coming out as a 64 char hexadecimal string. The passwords are compared in the login.php file method