HUGE PUSH - HYPE URL UPDATE
Embedded php to update the project list on the dashboard. Now each project acts as a link to a page with this url model: 'localhost/php/projects.php?projectID'. The projectID is a url parameter and the reason is to be able to access the projectID while in pages associated with that project(i.e. project dashboard, file creation page for that project). For now the project dashboard only displays the project ID, but we will create a comprehensive project dashboard in future commits to this branch.